requirements.txt
streamlit>=1.23.0
requests>=2.28.0
ratelimit>=2.2.1
pandas
google-generativeai
logging